% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_all.R
\name{collect_all}
\alias{collect_all}
\title{Performs the full analysis on all the .RDa files present in a
given folder. These .RDa files are in the format
created by the 'wirittttes' package}
\usage{
collect_all(folder = "~/GitHubs/wiritttea/inst/extdata")
}
\arguments{
\item{folder}{path of the folder containing the RDa files}
}
\value{
nothing, but will create multiple .CSV files in the folder
}
\description{
Performs the full analysis on all the .RDa files present in a
given folder. These .RDa files are in the format
created by the 'wirittttes' package
}
\examples{
  folder <- "~/GitHubs/wiritttea/inst/extdata"
  if (file.exists(folder)) {
    do_analysis(folder = folder)
    testit::assert(length(list.files(folder, pattern = "\\\\.csv")) > 0)
  }
}

