% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_n_alignments.R
\name{collect_file_n_alignments}
\alias{collect_file_n_alignments}
\title{Collect the number of alignments in a parameter file}
\usage{
collect_file_n_alignments(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the number of alignments in a parameter file
}
\examples{
 filename <- find_path("toy_example_3.RDa")
 df <- collect_file_n_alignments(filename)
 testit::assert(names(df) == c("n_alignments"))
 testit::assert(ncol(df) == 1)
 testit::assert(nrow(df) == 1)
 testit::assert(df$n_alignments[1] == 4)
}
\author{
Richel Bilderbeek
}

