% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_nltt_stats.R
\name{collect_file_nltt_stats}
\alias{collect_file_nltt_stats}
\title{Calculates the nLTT statistics from a file}
\usage{
collect_file_nltt_stats(filename)
}
\arguments{
\item{filename}{the name of a file}
}
\value{
a distribution of nLTT statistics
}
\description{
Calculates the nLTT statistics from a file
}
\examples{
  nltt_stats <- collect_file_nltt_stats(
    filename = find_path("toy_example_1.RDa")
  )
  expected_names <- c("sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(names(nltt_stats) == expected_names)
}
\author{
Richel Bilderbeek
}
