% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_nrbss.R
\name{collect_file_nrbss}
\alias{collect_file_nrbss}
\title{Collects the NRBS values between species trees and
posteriors in a parameter file in the melted/uncast/long form}
\usage{
collect_file_nrbss(filename)
}
\arguments{
\item{filename}{name of the parameter file}
}
\value{
A dataframe of NRBS values between each species tree and their posteriors
}
\description{
Collects the NRBS values between species trees and
posteriors in a parameter file in the melted/uncast/long form
}
\examples{
  filename <- find_path("toy_example_3.RDa")
  df <- collect_file_nrbss(filename)
  testit::assert(names(df) == c("sti", "ai", "pi", "si", "nrbs"))
  testit::assert(nrow(df) == 80)
}

