% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_posterior_crown_ages.R
\name{collect_file_posterior_crown_ages}
\alias{collect_file_posterior_crown_ages}
\title{Collect the crown ages of the BEAST2 posteriors}
\usage{
collect_file_posterior_crown_ages(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the crown ages of the BEAST2 posteriors
}
\note{
BEAST2 calls this parameter 'TreeHeight'
}
\examples{
  filename <- wiritttea::find_path("toy_example_3.RDa")
  df <- wiritttea::collect_file_posterior_crown_ages(filename)
  testthat::expect_equal(
    names(df),
    c("sti", "ai", "pi", "si", "crown_age")
  )
  testthat::expect_true(nrow(df) == 88)
}
\author{
Richel Bilderbeek
}
