% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_posterior_likelihoods.R
\name{collect_file_pstr_likelihoods}
\alias{collect_file_pstr_likelihoods}
\title{Collect the tree likelihoods of the BEAST2 posteriors}
\usage{
collect_file_pstr_likelihoods(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the tree likelihoods of the BEAST2 posteriors
}
\author{
Richel Bilderbeek
}
