% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_nltt_stats.R
\name{collect_files_nltt_stats}
\alias{collect_files_nltt_stats}
\title{Calculates the nLTT statistics from multiple clean files}
\usage{
collect_files_nltt_stats(filenames)
}
\arguments{
\item{filenames}{the names of the files}
}
\value{
a distribution of nLTT statistics
}
\description{
Calculates the nLTT statistics from multiple clean files
}
\examples{
  nltt_stats <- wiritttea::collect_files_nltt_stats(
    filenames = c(
      find_path("toy_example_1.RDa"),
      find_path("toy_example_2.RDa")
    )
  )
  expected_names <- c("filename", "sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(all(names(nltt_stats) == expected_names))
}
\author{
Richel Bilderbeek
}
