% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_nltt_stats.R
\name{collect_files_nltt_stats_dirty}
\alias{collect_files_nltt_stats_dirty}
\title{Calculates the nLTT statistics from multiple files that may be invalid}
\usage{
collect_files_nltt_stats_dirty(filenames)
}
\arguments{
\item{filenames}{the names of the files}
}
\value{
a distribution of nLTT statistics
}
\description{
Calculates the nLTT statistics from multiple files that may be invalid
}
\examples{
  nltt_stats <- collect_files_nltt_stats_dirty(
    filenames = c(
      find_path("toy_example_1.RDa"),
      "inva.lid",
      find_path("toy_example_2.RDa")
    )
  )
  expected_names <- c("filename", "sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(all(names(nltt_stats) == expected_names))
}
\author{
Richel Bilderbeek
}
