% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_nrbss.R
\name{collect_files_nrbss}
\alias{collect_files_nrbss}
\title{Collects the NRBS values of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_nrbss(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A dataframes of NRBS values
}
\description{
Collects the NRBS values of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- c(
    find_path("toy_example_3.RDa"),
    find_path("toy_example_4.RDa")
  )
  df <- collect_files_nrbss(filenames)
  expected <- c(
    "filename", "sti", "ai",
    "pi", "si", "nrbs"
  )
  testit::assert(names(df) == expected)
  testit::assert(nrow(df) == 160)
}

