% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_operators.R
\name{collect_files_operators}
\alias{collect_files_operators}
\title{Calculates the BEAST2 operators from multiple files}
\usage{

  collect_files_operators(filenames = rep(find_path("collect_files_operators.xml.state"),
  2))
}
\arguments{
\item{filenames}{the names of the .xml.state files}
}
\value{
a data frame with the operators' acceptances
}
\description{
Calculates the BEAST2 operators from multiple files
}
\examples{
  nltt_stats <- collect_files_operators(
    filenames = rep(find_path("collect_files_operators.xml.state"), 2)
  )
}
\author{
Richel Bilderbeek
}
