% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_posterior_gammas.R
\name{collect_files_posterior_gammas}
\alias{collect_files_posterior_gammas}
\title{Collects the gamma statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_posterior_gammas(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A list with two dataframes of gamma statistics
}
\description{
Collects the gamma statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
 filenames <- c(
   find_path("toy_example_3.RDa"),
   find_path("toy_example_4.RDa")
 )
 df <- collect_files_posterior_gammas(filenames)
 testit::assert(
   names(df) ==
   c("filename", "sti", "gamma_stat")
 )
 testit::assert(nrow(df) == 160)
}

