% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_posterior_nltts.R
\name{collect_files_posterior_nltts}
\alias{collect_files_posterior_nltts}
\title{Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_posterior_nltts(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A dataframe of nLTTs
}
\description{
Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- wiritttea::find_paths(
     c("toy_example_1.RDa", "toy_example_2.RDa",
     "toy_example_3.RDa", "toy_example_4.RDa")
   )
   df <- wiritttea::collect_files_posterior_nltts(filenames)

   testit::assert(
     dplyr::all_equal(
       names(df),
       c("filename", "sti", "ai", "pi", "si", "nltt")
     )
   )
   testthat::expect_true(nrow(df) == 200)
}
