% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_posterior_crown_ages.R
\name{collect_files_pstr_crown_ages}
\alias{collect_files_pstr_crown_ages}
\title{Collects the tree crown_ages of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_pstr_crown_ages(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{shows progress if TRUE}
}
\value{
A data frame with filename, sti, ai, pi, si, crown_age
}
\description{
Collects the tree crown_ages of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- wiritttea::find_paths(
   c("toy_example_1.RDa", "toy_example_2.RDa",
   "toy_example_3.RDa", "toy_example_4.RDa")
 )
 df <- wiritttea::collect_files_pstr_crown_ages(filenames)
 testthat::expect_equal(
   names(df),
   c("filename", "sti", "ai", "pi", "si", "crown_age")
 )
 testthat::expect_true(nrow(df) == 220)
}
