% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_species_tree_nltts.R
\name{collect_files_species_tree_nltt_stats}
\alias{collect_files_species_tree_nltt_stats}
\title{Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_species_tree_nltt_stats(filenames, dt)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\value{
A list with two dataframes of nLTTs
}
\description{
Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}

