% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_strees_identical.R
\name{collect_files_strees_identical}
\alias{collect_files_strees_identical}
\title{Collects if the two sampled species trees (from the same incipient
species trees) have the same branching times.}
\usage{
collect_files_strees_identical(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A data frame with filenames and TRUE/FALSEs
}
\description{
Collects if the two sampled species trees (from the same incipient
species trees) have the same branching times.
}
\examples{
 filenames <- c(
   find_path("toy_example_3.RDa"),
   find_path("toy_example_4.RDa")
 )
 df <- collect_files_strees_identical(filenames)
 testit::assert(
   names(df) ==
   c("filename", "strees_identical")
 )
 testit::assert(nrow(df) == 2)
}

