% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_log_file_info.R
\name{collect_log_file_info}
\alias{collect_log_file_info}
\title{Collects information of a log file in the melted/uncast/long form}
\usage{
collect_log_file_info(filename)
}
\arguments{
\item{filename}{name of the .log file}
}
\value{
A dataframe of log file info
}
\description{
Collects information of a log file in the melted/uncast/long form
}
\examples{
  filename <- wiritttea::find_path("add_alignments_ok.log")
  df <- wiritttea::collect_log_file_info(filename)
  testit::assert(names(df) == c("filename", "exit_status"))
  testit::assert(nrow(df) == 1)
}
