% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_acceptance_mcmc_operators.R
\name{create_figure_acceptance_mcmc_operators}
\alias{create_figure_acceptance_mcmc_operators}
\title{Create the figure 'Operator acceptances'}
\usage{
create_figure_acceptance_mcmc_operators(operators, filename)
}
\arguments{
\item{operators}{data frame with the collected operator acceptances}

\item{filename}{the name of the file the figure is saved to}
}
\value{
nothing, it saves a file
}
\description{
Create the figure 'Operator acceptances'
}
\author{
Richel Bilderbeek
}
