% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_alignment_qualities.R
\name{create_figure_alignment_qualities}
\alias{create_figure_alignment_qualities}
\title{Create figure 'alignment_qualities'}
\usage{
create_figure_alignment_qualities(alignments, filename)
}
\arguments{
\item{alignments}{data frame with alignments, as returned by 'read_collected_alignments'}

\item{filename}{the name of the file the figure is saved to}
}
\value{
nothing, it saves a file
}
\description{
Create figure 'alignment_qualities'
}
\author{
Richel Bilderbeek
}
