% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error.R
\name{create_figure_error}
\alias{create_figure_error}
\title{Creates figures 'figure_error.svg', 'figure_error_head.svg'
and 'figure_error_tail.svg', showing the complete distribution of all
errors (measured as nLTT statistic)}
\usage{
create_figure_error(parameters, nltt_stats, filename, verbose = FALSE)
}
\arguments{
\item{parameters}{collected parameters, as returned from 'collect_parameters'}

\item{nltt_stats}{collected nLTT statistics, as returned from
'read_collected_nltt_stats', assumes burn-in is already removed}

\item{verbose}{if set to TRUE, the function prints more information}

\item{svg_filenames}{the three filenames the figures will be saved as}
}
\description{
Creates figures 'figure_error.svg', 'figure_error_head.svg'
and 'figure_error_tail.svg', showing the complete distribution of all
errors (measured as nLTT statistic)
}
\examples{
  parameters <- read_collected_parameters()
  nltt_stats <- read_collected_nltt_stats(burn_in = 0.2)
  create_figure_error(
    parameters = parameters,
    nltt_stats = nltt_stats,
    filenames = paste0(path.expand("~"), "/figure_error_.svg")
  )
}
\author{
Richel Bilderbeek
}
