% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_alignment_length.R
\name{create_figure_error_alignment_length}
\alias{create_figure_error_alignment_length}
\title{Create figure 'error_alignment_length'}
\usage{
create_figure_error_alignment_length(parameters, nltt_stats, filename,
  n_sampled = nrow(stats::na.omit(nltt_stats)))
}
\arguments{
\item{parameters}{parameters, as returned by read_collected_parameters}

\item{nltt_stats}{nLTT stats, as returned by read_collected_nltt_stats}

\item{filename}{name of the file the figure is saved to}

\item{the}{number of nLTT statistics sampled}
}
\value{
nothing, only creates a file
}
\description{
Create figure 'error_alignment_length'
}
\author{
Richel Bilderbeek
}
