% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_alignment_length_median.R
\name{create_figure_error_alignment_length_median}
\alias{create_figure_error_alignment_length_median}
\title{Create figure 'figure_error_alignment_length_median'}
\usage{
create_figure_error_alignment_length_median(parameters, nltt_stats, esses,
  filename)
}
\arguments{
\item{parameters}{parameters, as returned by read_collected_parameters}

\item{nltt_stats}{nLTT statistics, as returned by read_collected_nltt_stats}

\item{esses}{ESSes, as returned by read_collected_esses}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure 'figure_error_alignment_length_median'
}
\author{
Richel Bilderbeek
}
