% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_bd_mean.R
\name{create_figure_error_bd_mean}
\alias{create_figure_error_bd_mean}
\title{Create figure create_figure_error_bd_mean}
\usage{
create_figure_error_bd_mean(parameters, nltt_stats, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure create_figure_error_bd_mean
}
