% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_low_high_ess.R
\name{create_figure_error_low_high_ess}
\alias{create_figure_error_low_high_ess}
\title{Create figure 'figure_error_low_high_ess'}
\usage{
create_figure_error_low_high_ess(parameters, nltt_stats, esses, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{esses}{the ESSes, as returned from read_collected_esses}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure 'figure_error_low_high_ess'
}
