% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_max.R
\name{create_figure_error_max}
\alias{create_figure_error_max}
\title{Create 'figure_error_max', showing the phylogeny with the biggest
  nLTT statistics error
  of all simulations}
\usage{
create_figure_error_max(parameters, nltt_stats, filename, sample_size = NA,
  raw_data_path = system.file("extdata", package = "wiritttea"))
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from
read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}

\item{sample_size}{the number of nLTT statistics that will be sampled, use
NA to sample all}

\item{raw_data_path}{the path where the raw data (the .RDa files) can
be found}
}
\description{
Create 'figure_error_max', showing the phylogeny with the biggest
  nLTT statistics error
  of all simulations
}
\author{
Richel Bilderbeek
}
