% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_figure_error_mean_dur_spec_alignment_length.R
\name{create_figure_error_mean_dur_spec_alignment_length}
\alias{create_figure_error_mean_dur_spec_alignment_length}
\title{Create 'figure_error_expected_mean_dur_spec_alignment_length'}
\usage{
create_figure_error_mean_dur_spec_alignment_length(parameters, nltt_stats,
  filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_error_expected_mean_dur_spec_alignment_length'
}
