% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_posterior_nltt_si.R
\name{create_figure_error_posterior_nltt_si}
\alias{create_figure_error_posterior_nltt_si}
\title{Create 'figure_error_posterior_nltt_si',
 showing phylogenies with the top 10 biggest nLTT statistics error
of all simulations}
\usage{
create_figure_error_posterior_nltt_si(parameters, nltt_stats, filename,
  sample_size = NA)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}

\item{sample_size}{the number of nLTT statistics that will be sampled, use
NA to sample all}
}
\description{
Create 'figure_error_posterior_nltt_si',
 showing phylogenies with the top 10 biggest nLTT statistics error
of all simulations
}
\author{
Richel Bilderbeek
}
