% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_sampling_representative.R
\name{create_figure_error_sampling_representative}
\alias{create_figure_error_sampling_representative}
\title{Create figure figure_error_sampling_representative}
\usage{
create_figure_error_sampling_representative(parameters, nltt_stats, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure figure_error_sampling_representative
}
