% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_tree_size.R
\name{create_figure_error_tree_size}
\alias{create_figure_error_tree_size}
\title{Create figure 'figure_error_tree_size'}
\usage{
create_figure_error_tree_size(n_taxa, nltt_stats, filename)
}
\arguments{
\item{n_taxa}{the number of taxa, as returned from read_collected_n_taxa}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure 'figure_error_tree_size'
}
