% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_ess_distribution.R
\name{create_figure_ess_distribution}
\alias{create_figure_ess_distribution}
\title{Create 'figure_ess_distribution'}
\usage{
create_figure_ess_distribution(parameters, df_esses, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{filename}{name of the file the figure will be saved to}

\item{esses}{the ESSes, as returned from read_collected_esses}
}
\description{
Create 'figure_ess_distribution'
}
