% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_ess_distribution_likelihood.R
\name{create_figure_ess_distribution_likelihood}
\alias{create_figure_ess_distribution_likelihood}
\title{Create 'figure_ess_distribution_likelihood'}
\usage{
create_figure_ess_distribution_likelihood(parameters, esses, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{esses}{the ESSes, as returned from read_collected_esses}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_ess_distribution_likelihood'
}
\author{
Richel Bilderbeek
}
