% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_exit_statuses.R
\name{create_figure_exit_statuses}
\alias{create_figure_exit_statuses}
\title{Create figure 'figure_exit_statuses'}
\usage{
create_figure_exit_statuses(log_files_info, filename)
}
\arguments{
\item{log_files_info}{log files' info, as returned from read_collected_log_files_info}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure 'figure_exit_statuses'
}
\author{
Richel J.C. Bilderbeek
}
