% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_exit_statuses_posteriors.R
\name{create_figure_exit_statuses_posteriors}
\alias{create_figure_exit_statuses_posteriors}
\title{Create figure_exit_statuses_posteriors: How many parameter estimates are OK?}
\usage{
create_figure_exit_statuses_posteriors(esses, filename)
}
\arguments{
\item{esses}{the ESSes, as returned from read_collected_esses}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure_exit_statuses_posteriors: How many parameter estimates are OK?
}
