% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_lowest_ess.R
\name{create_figure_lowest_ess}
\alias{create_figure_lowest_ess}
\title{Create figure_lowest_ess: Which estimated parameter has the lowest ESS?}
\usage{
create_figure_lowest_ess(esses, filename)
}
\arguments{
\item{esses}{the ESSes, as returned from read_collected_esses}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure_lowest_ess: Which estimated parameter has the lowest ESS?
}
