% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_posterior_bd.R
\name{create_figure_posterior_bd}
\alias{create_figure_posterior_bd}
\title{Create 'figure_posterior_distributions_bd'}
\usage{
create_figure_posterior_bd(parameters, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_posterior_distributions_bd'
}
