% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_posterior_crown_ages.R
\name{create_figure_posterior_crown_ages}
\alias{create_figure_posterior_crown_ages}
\title{Create 'figure_posterior_distributions_crown_age'}
\usage{
create_figure_posterior_crown_ages(posterior_crown_ages, filename)
}
\arguments{
\item{posterior_crown_ages}{posterior crown ages, as returned from read_collected_pstr_crown_ages}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_posterior_distributions_crown_age'
}
