% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_posterior_likelihoods.R
\name{create_figure_posterior_likelihoods}
\alias{create_figure_posterior_likelihoods}
\title{Create 'figure_posterior_distributions_likelihood' and 'figure_posterior_distributions_nltt'}
\usage{
create_figure_posterior_likelihoods(posterior_likelihoods, filename)
}
\arguments{
\item{posterior_likelihoods}{posterior likelihoods, as returned from read_collected_posterior_likelihoods}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_posterior_distributions_likelihood' and 'figure_posterior_distributions_nltt'
}
