% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_posterior_nltts.R
\name{create_figure_posterior_nltts}
\alias{create_figure_posterior_nltts}
\title{Create 'figure_posterior_distributions_nltts'}
\usage{
create_figure_posterior_nltts(nltt_stats, filename)
}
\arguments{
\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_posterior_distributions_nltts'
}
