% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_run_times.R
\name{create_figure_run_times}
\alias{create_figure_run_times}
\title{figure_run_times: distribution of runtimes}
\usage{
create_figure_run_times(log_files_info, filename)
}
\arguments{
\item{log_files_info}{log files' info, as returned from read_collected_log_files_info}

\item{filename}{name of the file the figure will be saved to}
}
\description{
figure_run_times: distribution of runtimes
}
