% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_strees_different.R
\name{create_figure_strees_different}
\alias{create_figure_strees_different}
\title{Create figure_strees_different
Analyse if the two sampled species trees are identical}
\usage{
create_figure_strees_different(strees_identical, filename)
}
\arguments{
\item{strees_identical}{dataframe with which species trees are identical,
as returned from read_collected_strees_identical}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create figure_strees_different
Analyse if the two sampled species trees are identical
}
