% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_109.R
\name{create_table_109}
\alias{create_table_109}
\title{Table 109: Successes and reasons of failures}
\usage{
create_table_109(log_files_info, filename)
}
\arguments{
\item{log_files_info}{log files' info, as returned from read_collected_log_files_info}

\item{filename}{name of the file the table will be saved to}
}
\description{
Table 109: Successes and reasons of failures
}
