% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_111.R
\name{create_table_111}
\alias{create_table_111}
\title{Table 111: How many parameter estimates are OK?}
\usage{
create_table_111(esses, filename)
}
\arguments{
\item{esses}{the ESSes, as returned from read_collected_esses}

\item{filename}{name of the file the table will be saved to}
}
\description{
Table 111: How many parameter estimates are OK?
}
