% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_112.R
\name{create_table_112}
\alias{create_table_112}
\title{Table 112: Which estimated parameter has the lowest ESS?}
\usage{
create_table_112(df_esses, filename)
}
\arguments{
\item{filename}{name of the file the table will be saved to}

\item{esses}{the ESSes, as returned from read_collected_esses}
}
\description{
Table 112: Which estimated parameter has the lowest ESS?
}
