% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_analysis.R
\name{do_analysis}
\alias{do_analysis}
\title{Performs the full analysis on all the .RDa files present in a
given folder. These .RDa files are in the format
created by the 'wirittttes' package}
\usage{
do_analysis(folder = "~/GitHubs/wiritttea/inst/extdata")
}
\arguments{
\item{folder}{path of the folder containing the RDa files}
}
\value{
nothing
}
\description{
Performs the full analysis on all the .RDa files present in a
given folder. These .RDa files are in the format
created by the 'wirittttes' package
}
\examples{
  folder <- system.file("extdata", package = "wiritttea")
  do_analysis(folder)
}

