% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_lines.R
\name{file_to_lines}
\alias{file_to_lines}
\title{Read a file and convert it to a vector of strings, in complexity O(2n)}
\usage{
file_to_lines(filename)
}
\arguments{
\item{filename}{the name of a file}
}
\value{
the text of the filename as a vector
}
\description{
Read a file and convert it to a vector of strings, in complexity O(2n)
}
\examples{
  text <- wiritttea::file_to_lines(
    filename = wiritttea::find_path("add_alignments_ok.log"))
  testit::assert(length(text) == 50)
}
\author{
Richel Bilderbeek
}
