% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_good_alignment.R
\name{is_good_alignment}
\alias{is_good_alignment}
\title{Checks if an alignment is good enough}
\usage{
is_good_alignment(alignment)
}
\arguments{
\item{alignment}{A DNA alignment, assumed to be created by the
Jukes-Cantor 1969 nucleotide substitution model}
}
\value{
TRUE or FALSE
}
\description{
Checks if an alignment is good enough
}
\author{
Richel Bilderbeek
}

