% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_alignments.R
\name{read_collected_alignments}
\alias{read_collected_alignments}
\title{Reads the file with the collected alignments' qualities}
\usage{
read_collected_alignments(csv_filename)
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all alignments' qualities
}
\description{
Reads the file with the collected alignments' qualities
}
\author{
Richel Bilderbeek
}
