% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_gammas_postrs.R
\name{read_collected_gammas_postrs}
\alias{read_collected_gammas_postrs}
\title{Read all the collected gamma statistics of all posteriors}
\usage{
read_collected_gammas_postrs(
  filename = wiritttea::find_path("collect_files_posterior_gammas.csv"))
}
\arguments{
\item{filename}{name of the CSV file}
}
\value{
a dataframe
}
\description{
Read all the collected gamma statistics of all posteriors
}
\examples{
  df <- read_collected_gammas_postrs()
  testit::assert(names(df) ==
    c(
      "filename", "sti", "ai",
      "pi", "si", "gamma_stat"
    )
  )
}
\author{
Richel Bilderbeek
}
