% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_gammas_strees.R
\name{read_collected_gammas_strees}
\alias{read_collected_gammas_strees}
\title{Read all the collected nLTT values of all species trees}
\usage{
read_collected_gammas_strees()
}
\arguments{
\item{filename}{name of the CSV file}
}
\value{
a dataframe
}
\description{
Read all the collected nLTT values of all species trees
}
\examples{
  df <- read_collected_gammas_strees()
  testit::assert(names(df) == c("filename", "sti", "gamma_stat"))
}
\author{
Richel Bilderbeek
}

