% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_log_files_info.R
\name{read_collected_log_files_info}
\alias{read_collected_log_files_info}
\title{Reads the file with the collected log files info}
\usage{
read_collected_log_files_info(csv_filename)
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all log files' information
}
\description{
Reads the file with the collected log files info
}
\author{
Richel Bilderbeek
}
