% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_n_taxa.R
\name{read_collected_n_taxa}
\alias{read_collected_n_taxa}
\title{Reads the file with the collected number of taxa}
\usage{

  read_collected_n_taxa(csv_filename = wiritttea::find_path("collect_files_n_taxa.csv"))
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all parameters
}
\description{
Reads the file with the collected number of taxa
}
\author{
Richel Bilderbeek
}

