% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_nltt_stats.R
\name{read_collected_nltt_stats}
\alias{read_collected_nltt_stats}
\title{Read all the collected nLTT statistics of all simulations}
\usage{

  read_collected_nltt_stats(filename = find_path("collect_files_nltt_stats.csv"))
}
\arguments{
\item{filename}{the name of the CSV containing the collected nLTT
statistics. Default value is 'inst/extdata/collected_files_nltt_stats.csv'}
}
\value{
a dataframe, with columns `filename` (name of the file),
  `sti` (species tree index), `ai` (alignment index), `pi` (posterior
  index), `si` (state index) and `nltt_stat` (the nLTT statistic between
  that posteriors's state and the nLTT of the original
  species tree (which is either 'youngest' for `sti` equals 1, or
  'oldest' for `sti` equals 2)
}
\description{
Read all the collected nLTT statistics of all simulations
}
\examples{
  df <- read_collected_nltt_stats()
  expected_names <- c("filename", "sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(names(df) == expected_names)
  testit::assert(is.factor(df$filename))
  testit::assert(is.factor(df$sti))
  testit::assert(is.factor(df$ai))
  testit::assert(is.factor(df$pi))
  testit::assert(is.factor(df$si))
}
\author{
Richel Bilderbeek
}
