% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_gammas_strees.R, R/read_collected_nltts_strees.R
\name{read_collected_nltts_strees}
\alias{read_collected_nltts_strees}
\title{Read all the collected nLTT values of all species trees}
\usage{
read_collected_nltts_strees()

read_collected_nltts_strees()
}
\value{
a dataframe

a dataframe
}
\description{
Read all the collected nLTT values of all species trees

Read all the collected nLTT values of all species trees
}
\examples{
  df <- read_collected_nltts_strees()
  testit::assert(names(df) == c("filename", "sti", "t", "nltt"))
  df <- read_collected_nltts_strees()
  testit::assert(names(df) == c("filename", "sti", "t", "nltt"))
}
\author{
Richel Bilderbeek

Richel Bilderbeek
}

