% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_operators.R
\name{read_collected_operators}
\alias{read_collected_operators}
\title{Reads the file with the collected BEAST2 operators' acceptances}
\usage{

  read_collected_operators(csv_filename = wiritttea::find_path("read_collected_operators.csv"))
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all operators' acceptances
}
\description{
Reads the file with the collected BEAST2 operators' acceptances
}
\author{
Richel Bilderbeek
}
