% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_parameters.R
\name{read_collected_parameters}
\alias{read_collected_parameters}
\title{Reads the file with the collected parameters}
\usage{
read_collected_parameters(
  csv_filename = wiritttea::find_path("collect_files_parameters.csv"))
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all parameters
}
\description{
Reads the file with the collected parameters
}
\examples{
  csv_filename <- wiritttea::find_path("collect_files_parameters.csv")
  df <- wiritttea::read_collected_parameters(csv_filename)
  testit::assert(nrow(df) == 4)
  testit::assert("filename" \%in\% names(df))
  testit::assert("rng_seed" \%in\% names(df))
  testit::assert("sirg" \%in\% names(df))
  testit::assert("siri" \%in\% names(df))
  testit::assert("scr" \%in\% names(df))
  testit::assert("erg" \%in\% names(df))
  testit::assert("eri" \%in\% names(df))
  testit::assert("age" \%in\% names(df))
  testit::assert("mutation_rate" \%in\% names(df))
  testit::assert("n_alignments" \%in\% names(df))
  testit::assert("sequence_length" \%in\% names(df))
  testit::assert("nspp" \%in\% names(df))
  testit::assert("n_beast_runs" \%in\% names(df))
}
\author{
Richel Bilderbeek
}
