% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_postr_cas.R
\name{read_collected_posterior_crown_ages}
\alias{read_collected_posterior_crown_ages}
\title{Read all the collected tree crown_ages of all posteriors}
\usage{
read_collected_posterior_crown_ages(
  filename ="collected_posterior_crown_ages.csv"
)
}
\arguments{
\item{filename}{name of the file with all the collected posteriors' crown ages}
}
\value{
a dataframe
}
\description{
Read all the collected tree crown_ages of all posteriors
}
\examples{
  df <- wiritttea::read_collected_posterior_crown_ages()
  testit::assert(
    all(
      names(df) ==  c("filename", "sti", "ai", "pi", "si", "crown_age")
    )
  )
}
\author{
Richel Bilderbeek
}
