% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_strees_identical.R
\name{read_collected_strees_identical}
\alias{read_collected_strees_identical}
\title{Reads the file with the collected species trees being identical}
\usage{

  read_collected_strees_identical(csv_filename = wiritttea::find_path("collect_files_strees_identical.csv"))
}
\arguments{
\item{csv_filename}{name of the CSV file}
}
\value{
a data frame with all parameters
}
\description{
Reads the file with the collected species trees being identical
}
\examples{
  csv_filename <- wiritttea::find_path("collect_files_strees_identical.csv")
  strees_identical <- read_collected_strees_identical(csv_filename)
  testit::assert(names(df) == c(
    "filename", "strees_identical"
    )
  )
  testit::assert(is.factor(df$filename))
  testit::assert(is.factor(df$strees_identical))
}
\author{
Richel Bilderbeek
}
