% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_alignments.R
\name{collect_files_alignments}
\alias{collect_files_alignments}
\title{Collects the quality of the alignments of the files}
\usage{
collect_files_alignments(my_filenames)
}
\arguments{
\item{filenames}{names of the parameter files}
}
\value{
A dataframe with the columns 'filename', 'n_alignments_ok' (number
  of alignments that have a distance matrix with only non-zero and non-NA
  elements, except on the diagonal), 'n_alignments_zero' (number
  of alignments that have a distance matrix with zeroes next to the
  diagonal, indicating two identical DNA sequences in the alignment),
  and 'n_alignments_na' (number
  of alignments that have a distance matrix with NAs,
  indicating that two DNA sequences are two dissimilar to have
  their Jukes-Cantor distance measured)
}
\description{
Collects the quality of the alignments of the files
}
\examples{
  filenames <- find_paths(c("toy_example_3.RDa", "toy_example_4.RDa"))
  df <- collect_files_alignments(filenames)
  testit::assert(nrow(df) == 2)
  expected_names <- c(
    "filename",
    "n_alignments_ok",
    "n_alignments_zeroes",
    "n_alignments_na")
  testit::assert(all.equal(names(df), expected_names))
}
