% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_posterior_likelihoods.R
\name{collect_files_posterior_likelihoods}
\alias{collect_files_posterior_likelihoods}
\title{Collects the tree likelihoods of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_posterior_likelihoods(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A data frame with filename, sti, ai, pi, si, likelihood
}
\description{
Collects the tree likelihoods of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
