% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_log_files_info.R
\name{collect_log_files_info}
\alias{collect_log_files_info}
\title{Collects log file information in the melted/uncast/long form}
\usage{
collect_log_files_info(filenames)
}
\arguments{
\item{filenames}{names of the parameter file}
}
\value{
A dataframe with all information of all log files
}
\description{
Collects log file information in the melted/uncast/long form
}
\examples{
  filenames <- c(
   find_path("add_alignments_ok.log"),
   find_path("add_alignments_exceeded_memory.log")
 )
 df <- collect_log_files_info(filenames)
 testit::assert(names(df) == c("filename", "exit_status"))
 testit::assert(nrow(df) == length(filenames))
 testit::assert(df$exit_status[1] != df$exit_status[2])
}
