% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error.R
\name{create_figure_error}
\alias{create_figure_error}
\title{Creates figures 'figure_error.svg', 'figure_error_head.svg'
and 'figure_error_tail.svg', showing the complete distribution of all
errors (measured as nLTT statistic)}
\usage{
create_figure_error(parameters, nltt_stats, cut_x = 0.05,
  path = path.expand("~"), verbose = FALSE)
}
\arguments{
\item{parameters}{collected parameters, as returned from 'collect_parameters'}

\item{nltt_stats}{collected nLTT statistics, as returned from
'read_collected_nltt_stats', assumes burn-in is already removed}

\item{cut_x}{nLTT statistic value at which head and tail are seperated}

\item{path}{the path of where the figures will be saved, which is the
home folder by default}

\item{verbose}{if set to TRUE, the function prints more information}
}
\description{
Creates figures 'figure_error.svg', 'figure_error_head.svg'
and 'figure_error_tail.svg', showing the complete distribution of all
errors (measured as nLTT statistic)
}
\examples{
  parameters <- read_collected_parameters()
  nltt_stats <- read_collected_nltt_stats(burn_in = 0.2)
  create_figure_error(
    parameters = parameters,
    nltt_stats = nltt_stats,
    path = path.expand("~")
  )
}
\author{
Richel Bilderbeek
}
