% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure_error_max.R
\name{create_figure_error_max}
\alias{create_figure_error_max}
\title{Create 'figure_error_max', showing phylogenies with the top 10 biggest nLTT statistics error
of all simulations}
\usage{
create_figure_error_max(parameters, nltt_stats, filename)
}
\arguments{
\item{parameters}{parameters, as returned from read_collected_parameters}

\item{nltt_stats}{the nLTT statistics, as returned from read_collected_nltt_stats}

\item{filename}{name of the file the figure will be saved to}
}
\description{
Create 'figure_error_max', showing phylogenies with the top 10 biggest nLTT statistics error
of all simulations
}
\author{
Richel Bilderbeek
}
