% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_postr_nltts.R
\name{read_collected_posterior_nltts}
\alias{read_collected_posterior_nltts}
\title{Read all the collected nLTT statistics of all posteriors}
\usage{
read_collected_posterior_nltts(
    filename = wiritttea::find_path("collect_files_posterior_nltts.csv"),
    burn_in_fraction
  )
}
\arguments{
\item{filename}{name of the file with all the collected posteriors' nLLTs}

\item{burn_in_fraction}{the fraction of earliest values that is discarded}
}
\value{
a dataframe
}
\description{
Read all the collected nLTT statistics of all posteriors
}
\examples{
  df <- wiritttea::read_collected_posterior_nltts(
    filename = wiritttea::find_path("collect_files_posterior_nltts.csv"),
    burn_in_fraction = 0.2
  )
  testit::assert(names(df) ==
    c(
      "filename", "sti", "ai",
      "pi", "si", "nltt"
    )
  )
}
\author{
Richel Bilderbeek
}
