% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_postr_nltts.R
\name{read_collected_posterior_nltts}
\alias{read_collected_posterior_nltts}
\title{Read all the collected nLTT statistics of all posteriors}
\usage{

  read_collected_posterior_nltts(filename = wiritttea::find_path("collect_files_posterior_nltts.csv"))
}
\arguments{
\item{filename}{name of the file with all the collected posteriors' nLLTs}
}
\value{
a dataframe
}
\description{
Read all the collected nLTT statistics of all posteriors
}
\examples{
  df <- read_collected_posterior_nltts()
  testit::assert(names(df) ==
    c(
      "filename", "sti", "ai",
      "pi", "si", "nltt"
    )
}
\author{
Richel Bilderbeek
}
